app/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
boto==2.39.0
celery==3.1.23
django-admin-sortable2==0.6.3
django-allauth==0.25.2
django-bootstrap3==7.0.1
django-ckeditor==5.0.3
django-compressor==2.0
django-debug-toolbar==1.4
django-flat-theme==1.1.3
django-haystack-panel==0.2.1
django-haystack==2.4.1
django-mptt==0.8.3
django-pure-pagination==0.3.0
django-json-rpc==0.7.2
django-smart-selects==1.2.2
django-storages==1.4.1
django-suit==0.2.18
Django==1.9.5
gunicorn==19.4.5
Jinja2==2.8
kombu==3.0.35
psycopg2==2.6
Pillow==3.2.0
pysolr==3.4.0
pytz==2016.3
redis==2.10.5
requests==2.9.1
simplejson==3.8.2
sorl-thumbnail==12.3
